requirements.txt
click==8.1.7
Jinja2==3.1.3
PyYAML==6.0.1
rich==13.7.1
tomli==2.0.1
